size matters

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,4 @@
-#include<stdio.h>
+#include <stdio.h>
 
 /**
  *main - execting function
@@ -14,11 +14,11 @@ int main(void)
 	long l;
 	long ll;
 
-	printf("size of a char: %lu byte\n", sizeof(character));
-	printf("size of an int: %lu byte(s)\n", size of(i));
-	printf("size of a float: %lu bytes\n", sizeof(f));
-	printf("size of a long: %lu bytes\n", sizeof(l));
-	printf("size of a long long: %lu bytes\n", sizeof(ll));
+	printf("size of a char: %lu byte\n", (unsigned long)sizeof(character));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)size of(i));
+	printf("size of a float: %lu bytes\n", (unsigned long)sizeof(f));
+	printf("size of a long: %lu bytes\n", (unsigned long)sizeof(l));
+	printf("size of a long long: %lu bytes\n", (unsigned long)sizeof(ll));
 
 	return (0);
 }
